Create Networks.py
I'm not worried about the formatting structure or any of that, I'll do a refactoring later down the line. I'd say this code is around 50% finished, all that's left is performing the actual forward propagation, looping this cycle in a genetic algorithm, and assessing the fitness of any of the solutions. Creating these networks ended up being a lot harder than I expected, mainly due to my complete lack of experience w/ LSTMs.
BTW Future Hudson: This code has a bug with the anet_wb and hnet_wb creation lines, specifically 110 and 112. Look into that!