Added in proper output for Anet and Hnet
The Architecture Network requires a specific output format, which uses a combination of multiple types of activation functions. Because of this, I also had to slightly re-design rand_dense_weights() to work along with the Functional API's concatenated layer. I also finished the propagation of data through the networks, and all that is left is decoding them, running them on the subnet (snet), and then integrating the genetic algorithm. The major parts of the code are nearly done.