Initial Version, Finished Model Structure
Created the whole network structure, I'm going with a single model design cause it makes so much more sense than the multiple model shenanigans of before, since with this I can reasonably create a custom training loop and use gradient descent to train it. I programmed this in VS Code, so the formatting might be a bit inconsistent from the usual Spyder formatting.