Create Dataset_Generator.py
The whole intent of this is to create some temporary test data to test the rest of the network on, just as a proof of concept before moving on to a larger set of data. Even if it just ends up overfitting, I want to know that the network has the capability to learn before I waste more time. This program uses the default Keras datasets.