Fully Functional Forward Propagation
The forward propagation for the neural network has been finished. This network can take in datasets and make predictions on them. Although it might look like a total overhaul of code, all I've really done is take the Functional API and implement it with Keras Model Subclassing, along with creating a custom layer for the layer. All that's left is finishing up the custom training loop and creating a way to evaluate loss. My estimate is that'll be ~2-3 more days of work. The first model should hopefully be running by next week.