Final Commit For Now...
Unfortunately, I have a lot of other stuff that I have to do, so this will probably be my final commit for now. I will hopefully continue this project when I have more free time. The only part left is finishing the VGG16 part, so it shouldn't be too far from completion.