Kaggle-Friendly Version of Optimized_Parameters.py
I have decided to run the program on Kaggle GPUs because it is estimated to take ~28 hours to run. Kaggle uses Python 3.7.12, so in order to make the program compatible, I've had to convert the match-cases into if-elifs. The Hyperband tuner takes ~13 minutes per model on my RTX 3060, and ~12.5 minutes per model on Kaggle GPUs. I have not modified it to work on TPUs yet. Also, in the previous commit I forgot to actually include the code, so that's why there are two identical commits.