Full refactor & added in generator
I fully refactored the program to work as 2 classes, and I also added in a new class that sequentially imports the dataset, which makes the possible training dataset size much much larger. This program still hasn't been fully tested, so I'd recommend analyzing the outputs of the generator later on to make sure it's properly working.